frontend_ETL/requirements.txt
appwrite>=7.0.0
python_dotenv>=0.0.5
requests>=2.32.3
beautifulsoup4>=4.12.3